backend/requirements.txt
onnxruntime-gpu
transformers
fastapi
uvicorn